// types.ts
export interface ClassHours {
  A?: number;
  B?: number;
  C?: number;
  [key: string]: number | undefined;
}

export interface TeacherData {
  id: number;
  no: string;
  name: string;
  nip?: string;
  rank?: string;
  gol?: string;
  subject: string;
  code: string;
  hoursVII: ClassHours;
  hoursVIII: ClassHours;
  hoursIX: ClassHours;
  additionalTask?: string;
  additionalHours: number;
  totalHours: number;
}